Test for existence of directory before attempting deletion

+++ RMS/ArchiveDetections.py
@@ -278,7 +278,9 @@ def archiveDetections(captured_path, archived_path, ff_detected, config, extra_f
         imgdata_archive_name = archive_base + "_imgdata"
         imgdata_archive_name = archiveDir(captured_path, imgdata_set, imgdata_archived_path,
                                           imgdata_archive_name, extra_files=extra_files)
-        shutil.rmtree(imgdata_archived_path)
+        if os.path.exists(imgdata_archived_path):
+            if os.path.isdir(imgdata_archived_path):
+                shutil.rmtree(imgdata_archived_path)
 
         # create a set which is file_list excluding all contents of imgdata_set
         metadata_set = set([item for item in file_list if item not in imgdata_set])
@@ -291,7 +293,9 @@ def archiveDetections(captured_path, archived_path, ff_detected, config, extra_f
 
         metadata_archive_name = archiveDir(captured_path, metadata_set, metadata_archived_path,
                                            metadata_archive_name, extra_files=extra_files)
-        shutil.rmtree(metadata_archived_path)
+        if os.path.exists(metadata_archived_path):
+            if os.path.isdir(metadata_archived_path):
+                shutil.rmtree(metadata_archived_path)
 
         return archive_name, imgdata_archive_name, metadata_archive_name
